+++ cartridges/int_klarna_payments/cartridge/scripts/common/klarnaPaymentsHttpService.js
@@ -84,6 +84,16 @@ KlarnaPaymentsHttpService.prototype.call = function( serviceID, urlPath, httpVer
     service.addHeader( 'Accept', 'application/json' );
     service.addHeader( 'User-Agent', SERVICE_HEADER );
 
+    // Add X-Klarna-Integration-Metadata header
+    var KlarnaConstants = require( '*/cartridge/scripts/util/klarnaPaymentsConstants' );
+    if ( KlarnaConstants.integrator && KlarnaConstants.originators ) {
+        var integrationMetadata = {
+            integrator: KlarnaConstants.integrator,
+            originators: KlarnaConstants.originators
+        };
+        service.addHeader( 'X-Klarna-Integration-Metadata', JSON.stringify( integrationMetadata ) );
+    }
+
     if ( !empty( klarnaIdempotencyKey ) && klarnaIdempotencyKey !== 'undefined' ) {
         service.addHeader( 'Klarna-Idempotency-Key', klarnaIdempotencyKey );
     }