+++ cartridges/int_klarna_payments_sfra/cartridge/client/default/js/klarnaSignIn.js
@@ -10,7 +10,9 @@ window.onload = async function () {
     const klarna = await Klarna.init({
         clientId: klarnaPreferences.kpSignInClientID,
         environment: klarnaPreferences.kpSignInEnvironment,
-        locale: klarnaPreferences.kpSignInLocale
+        locale: klarnaPreferences.kpSignInLocale,
+        integrator: klarnaPreferences.integrator,
+        originators: klarnaPreferences.originators
     });
     // if (!klarnaSDK.Identity.canMakeLogin('DE')) return;
     const siwkButton = klarna.Identity.button({